séparation de gems.py en 3 fichiers
