+++ get_gems.py
@@ -3,18 +3,18 @@ from discord.ext import commands, tasks
 from discord.ext.commands import Bot
 from discord.utils import get
 
-import DB
-import welcome as wel
+from DB import DB
+from core import welcome as wel
 
 # initialisation des variables.
 DEFAUT_PREFIX = "*"
 
 
-VERSION = open("fichier_txt/version.txt").read().replace("\n","")
-TOKEN = open("fichier_txt/token_getgems.txt", "r").read().replace("\n","")
+VERSION = open("core/version.txt").read().replace("\n","")
+TOKEN = open("token/token_getgems.txt", "r").read().replace("\n","")
 client = commands.Bot(command_prefix = "{0}".format(DEFAUT_PREFIX))
-NONE = open("fichier_txt/cogs.txt","w")
-NONE = open("fichier_txt/help.txt","w")
+NONE = open("help/cogs.txt","w")
+NONE = open("help/help.txt","w")
 
 client.remove_command("help")
 
@@ -34,15 +34,17 @@ async def on_member_remove(member):
 
 ####################### Commande help.py #######################
 
-client.load_extension('help')
+client.load_extension('help.help')
 
 ################### Core ############################
 
-client.load_extension('utils')
+client.load_extension('core.utils')
 
 ####################### Commande gems.py #######################
 
-client.load_extension('gems')
+client.load_extension('gems.gemsBase')
+
+client.load_extension('gems.gemsPlay')
 
 ####################### Lancemement du bot ######################
